[#12] CodiBoardView Component 분리시키기
#12
